Make `cargo test` compile again
Move benches into /benches directory.
Fix compiler warnings.
